requirements.txt
# Librerías necesarias para la Function App de Security Hub

azure-functions
boto3
azure-storage-blob